refactor: resolve 책임을 RequestBodyResolver로 이전
