+++ tomcat/src/main/java/org/apache/coyote/http11/controller/UserController.java
@@ -1,22 +1,17 @@
 package org.apache.coyote.http11.controller;
 
 import com.techcourse.model.User;
-import java.util.HashMap;
-import java.util.Map;
 import org.apache.coyote.http11.HttpMethod;
 import org.apache.coyote.http11.HttpStatusCode;
 import org.apache.coyote.http11.request.HttpRequest;
-import org.apache.coyote.http11.request.RequestBody;
+import org.apache.coyote.http11.request.resolver.RequestBodyResolver;
+import org.apache.coyote.http11.request.resolver.UserResolver;
 import org.apache.coyote.http11.response.HttpResponse;
 import org.apache.coyote.http11.service.UserService;
 
 public class UserController implements Controller {
 
-    private static final String USER_REGISTRATION_INFO_DELIMITER = "&";
-    private static final String INFO_ELEMENT_DELIMITER = "=";
-    private static final int ELEMENT_KEY_INDEX = 0;
-    private static final int ELEMENT_VALUE_INDEX = 1;
-
+    private final RequestBodyResolver<User> userResolver = new UserResolver();
     private final UserService userService = UserService.getInstance();
 
     @Override
@@ -27,7 +22,7 @@ public class UserController implements Controller {
     @Override
     public void service(HttpRequest httpRequest, HttpResponse httpResponse) {
         if (httpRequest.isMethod(HttpMethod.POST)) {
-            User user = resolveUser(httpRequest.getRequestBody());
+            User user = userResolver.resolve(httpRequest.getRequestBody());
             userService.save(user);
 
             httpResponse.statusCode(HttpStatusCode.FOUND_302)
@@ -37,22 +32,5 @@ public class UserController implements Controller {
 
         httpResponse.statusCode(HttpStatusCode.OK_200)
                 .viewUrl("/register.html");
-    }
-
-    private User resolveUser(RequestBody requestBody) {
-        Map<String, String> registerInfo = new HashMap<>();
-        String body = requestBody.getContent();
-        String[] elements = body.split(USER_REGISTRATION_INFO_DELIMITER);
-
-        for (String element : elements) {
-            String[] parsedElement = element.split(INFO_ELEMENT_DELIMITER);
-            registerInfo.put(parsedElement[ELEMENT_KEY_INDEX], parsedElement[ELEMENT_VALUE_INDEX]);
-        }
-
-        return new User(
-                registerInfo.get("account"),
-                registerInfo.get("password"),
-                registerInfo.get("email")
-        );
     }
 }
