+++ tomcat/src/main/java/org/apache/coyote/http11/request/resolver/RequestBodyResolver.java
@@ -1,0 +1,8 @@
+package org.apache.coyote.http11.request.resolver;
+
+import org.apache.coyote.http11.request.RequestBody;
+
+public interface RequestBodyResolver <T> {
+
+    T resolve(RequestBody requestBody);
+}
